refactor: made b.js the new index file
